Refaktorering av menyknapper
